+++ CHANGELOG.md
@@ -33,6 +33,8 @@
 
 - CHANGED: Updated whois.cmc.iq parser to the new response format.
 
+- CHANGED: Updated whois.na-nic.com.na parser to the new response format.
+
 - FIXED: whois.domainregistry.ie should support status `Active - LOCKED`.
 
 - FIXED: whois.nic.uk fails to parse registrars without URL (GH-188).